Move TokensService and -Repository into their own folder
