+++ src/app/authentication/authentication.module.ts
@@ -1,16 +1,15 @@
 import { Global, Module } from '@nestjs/common';
+import { TokensModule } from '../tokens';
 import { UsersModule } from '../users';
 import { AuthenticationController } from './authentication.controller';
 import { AuthenticationService } from './authentication.service';
 import { AuthorizationRepository } from './authorization.repository';
-import { TokensRepository } from './tokens.repository';
-import { TokensService } from './tokens.service';
 
 @Global()
 @Module({
-    imports: [UsersModule],
+    imports: [TokensModule, UsersModule],
     controllers: [AuthenticationController],
-    providers: [AuthenticationService, AuthorizationRepository, TokensService, TokensRepository],
-    exports: [AuthenticationService, TokensService],
+    providers: [AuthenticationService, AuthorizationRepository],
+    exports: [AuthenticationService],
 })
 export class AuthenticationModule {}
