+++ src/app/tokens/index.ts
@@ -1,0 +1,2 @@
+export * from './tokens.module';
+export * from './tokens.service';
